Change activity manager names, remove activity after iteration on period end
